requirements.txt
# Core GUI Framework
gradio>=4.0.0

# Image Processing and Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0  # Alternative option

# Data Manipulation and Numerical Computing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and File Handling
PyYAML>=6.0
configparser>=5.3.0

# Additional Utilities
tqdm>=4.65.0  # Progress bars for model loading/processing
pathlib2>=2.3.7  # Enhanced path handling
